Fixed Infinite Loop
Fixed Infinte Loop which was the product of inserting select variables
at every binary instruction. Since this also allowed to insert a select
variables into the loop latch of for-loop, the SAT solver generated a
placement which ended up in non-termination.

With this fix, no select variable will be inserted, iff the binary
statement is part of the loop-latch.